spot-checker, spot-ddd 동시실행을위한 태그 설정

+++ spot-availability-tester/aws/IaC/terminate-pending-instances.py
@@ -5,6 +5,7 @@ import json
 
 LOG_GROUP_NAME = os.environ['LOG_GROUP_NAME']
 LOG_STREAM_NAME = os.environ['LOG_STREAM_NAME']
+VPC_ID = os.environ['VPC_ID']
 
 ec2 = boto3.client('ec2')
 logs_client = boto3.client('logs')
@@ -20,14 +21,23 @@ def create_log_event(result):
 def lambda_handler(event, context):
     instances_to_terminate = []
     instances_to_terminate.append(event['detail']['instance-id'])
+    response = ec2.describe_instances(
+        Filters=[{'Name': 'instance-lifecycle', 'Values': ['spot']}],
+        InstanceIds=instances_to_terminate
+    )
 
-    log_data = {
-        "Timestamp": time.time(),
-        "InstanceId": event['detail']['instance-id'],
-        "InstanceState": event['detail']['state'],
-        "Region": event['region'],
-    }
-    create_log_event(json.dumps(log_data))
+    if not response['Reservations']:
+        return
 
-    ec2.terminate_instances(InstanceIds=instances_to_terminate)
-    print(f"Terminated instances: {event['detail']['instance-id']}")
+    instances_to_terminate_vpc = response['Reservations'][0]['Instances'][0]['VpcId']
+    if instances_to_terminate_vpc == VPC_ID:
+        log_data = {
+            "Timestamp": time.time(),
+            "InstanceId": event['detail']['instance-id'],
+            "InstanceState": event['detail']['state'],
+            "Region": event['region'],
+        }
+        create_log_event(json.dumps(log_data))
+
+        ec2.terminate_instances(InstanceIds=instances_to_terminate)
+        print(f"Terminated instances: {event['detail']['instance-id']}")
